last changes

+++ src/main/resources/liquibase/migrations/02_statistic.sql
@@ -6,7 +6,7 @@ CREATE TABLE statistic (
                            scope           VARCHAR(255) NOT NULL,
                            updated_at      TIMESTAMPTZ NOT NULL,
                            type            VARCHAR(255) NOT NULL,
-                           value           DOUBLE PRECISION NOT NULL,
+                           value           FLOAT NOT NULL,
                            task_id         UUID,
                            user_id         UUID,
                            event_type      VARCHAR(255) NOT NULL
